Gradio/requirements.txt
numpy
scipy
matplotlib
torchvision==0.2.1
nltk
Pillow
tqdm
lmdb
tensorflow
tensorboardX
